# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/auth/otp/sms": {
            "post": {
                "description": "Send a one-time password (OTP) to the provided phone number via SMS using Twilio Verify service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Send OTP via SMS",
                "parameters": [
                    {
                        "description": "Phone number request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/send_otp.Request"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/auth/otp/whatsapp": {
            "post": {
                "description": "Send a one-time password (OTP) to the provided phone number via WhatsApp using Twilio Verify service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Send OTP via WhatsApp",
                "parameters": [
                    {
                        "description": "Phone number request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/send_otp.Request"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/auth/verify-otp": {
            "post": {
                "description": "Verify the OTP code sent via SMS or WhatsApp, create user account if new, and return JWT access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Verify OTP and Authenticate User",
                "parameters": [
                    {
                        "description": "OTP verification request containing phone number and code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/verify_otp.Request"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/vendor/order-detail": {
            "get": {
                "description": "Retrieve comprehensive details about a specific order including delivery information, customer details, and all items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get Vendor Order Details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "order_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/vendors.OrderDetailsResponse"
                        }
                    }
                }
            }
        },
...
        "/api/vendor/profile": {
            "get": {
                "description": "Retrieve complete profile information for a specific vendor including business details, location, and operating hours",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vendor-profile"
                ],
                "summary": "Get Vendor Profile Details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "description": "Create a new vendor profile with complete business information, location details, and operating hours",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vendor-profile"
                ],
                "summary": "Create Vendor Profile",
                "parameters": [
                    {
                        "description": "Complete vendor profile information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vendors.CompleteVendorProfile"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/vendors/{vendor_id}/inventory": {
            "get": {
                "description": "Retrieve paginated inventory items for a specific vendor with filtering and search capabilities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Get Vendor Inventory Items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vendor ID",
                        "name": "vendor_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20, max: 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search items by name or description",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter items by stock status",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "description": "Add a new item from the master catalog to the vendor's inventory with initial stock quantity and optional price override",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Add Item to Vendor Inventory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vendor ID",
                        "name": "vendor_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Add item to inventory request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vendors.AddItemToInventoryRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/vendors/{vendor_id}/inventory/summary": {
            "get": {
                "description": "Retrieve summary statistics of vendor's inventory including total items, in-stock items, and out-of-stock items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Get Vendor Inventory Summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vendor ID",
                        "name": "vendor_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventory summary retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/vendors.InventorySummary"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "send_otp.Request": {
            "type": "object",
            "required": [
                "phone"
            ],
            "properties": {
                "phone": {
                    "description": "E.164 format: +1234567890",
                    "type": "string",
                    "example": "+918010201921"
                }
            }
        },
...
        "vendors.AddItemToInventoryRequest": {
            "type": "object",
            "required": [
                "item_id",
                "stock_quantity"
            ],
            "properties": {
                "item_id": {
                    "type": "integer",
                    "example": 456
                },
                "price_override": {
                    "type": "number",
                    "minimum": 0,
                    "example": 11.99
                },
                "stock_quantity": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 50
                }
            }
        },
...
        "vendors.CompleteVendorProfile": {
            "type": "object",
            "properties": {
                "account_type": {
                    "type": "string",
                    "enum": [
                        "basic",
                        "premium",
                        "enterprise"
                    ],
                    "example": "premium"
                },
                "address": {
                    "type": "string",
                    "example": "123 Main Street, Downtown, New York, NY 10001"
                },
                "business_name": {
                    "type": "string",
                    "example": "Smith's Fresh Market LLC"
                },
                "closing_time": {
                    "type": "string",
                    "example": "2000-01-01T21:00:00Z"
                },
                "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john.smith@freshmarket.com"
                },
                "gstin": {
                    "type": "string",
                    "example": "22AAAAA0000A1Z5"
                },
                "image_s3_url": {
                    "type": "string",
                    "example": "https://my-bucket.s3.amazonaws.com/vendors/store-123.jpg"
                },
                "latitude": {
                    "type": "number",
                    "maximum": 90,
                    "minimum": -90,
                    "example": 40.7128
                },
                "live_status": {
                    "type": "boolean",
                    "example": true
                },
                "longitude": {
                    "type": "number",
                    "maximum": 180,
                    "minimum": -180,
                    "example": -74.006
                },
                "opening_time": {
                    "type": "string",
                    "example": "2000-01-01T09:00:00Z"
                },
                "owner_name": {
                    "type": "string",
                    "example": "John Smith"
                },
                "phone": {
                    "type": "string",
                    "example": "+1234567890"
                }
            }
        },
        "vendors.InventorySummary": {
            "type": "object",
            "properties": {
                "in_stock_items": {
                    "type": "integer",
                    "example": 135
                },
                "out_of_stock_items": {
                    "type": "integer",
                    "example": 15
                },
                "total_items": {
                    "type": "integer",
                    "example": 150
                }
            }
        },
        "vendors.OrderDetailsResponse": {
            "type": "object",
            "properties": {
                "customer_address": {
                    "type": "string",
                    "example": "123 Main St, Apt 4B, New York, NY 10001"
                },
                "customer_name": {
                    "type": "string",
                    "example": "Jane Smith"
                },
                "delivered_by_time": {
                    "type": "string",
                    "example": "2025-05-30T15:00:00Z"
                },
                "delivery_partner_name": {
                    "type": "string",
                    "example": "John Delivery"
                },
                "delivery_partner_phone": {
                    "type": "string",
                    "example": "+1234567890"
                },
                "delivery_partner_pin": {
                    "type": "string",
                    "example": "1234"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vendors.OrderItem"
                    }
                },
                "order_created_time": {
                    "type": "string",
                    "example": "2025-05-30T13:30:00Z"
                },
                "order_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "order_status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "accepted",
                        "packed",
                        "ready",
                        "completed",
                        "cancelled",
                        "rejected"
                    ],
                    "example": "packed"
                },
                "order_total_amount": {
                    "type": "number",
                    "example": 45.99
                },
                "pack_by_time": {
                    "type": "string",
                    "example": "2025-05-30T14:30:00Z"
                },
                "paid_by_time": {
                    "type": "string",
                    "example": "2025-05-30T13:45:00Z"
                }
            }
        },
        "vendors.OrderItem": {
            "type": "object",
            "properties": {
                "item_id": {
                    "type": "integer",
                    "example": 1
                },
                "item_image_url": {
                    "type": "string",
                    "example": "https://example.com/images/margherita-pizza.jpg"
                },
                "item_name": {
                    "type": "string",
                    "example": "Margherita Pizza"
                },
                "qty_ordered": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 2
                }
            }
        },
...
        "verify_otp.Request": {
            "type": "object",
            "required": [
                "code",
                "device_info",
                "phone"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "device_info": {
                    "type": "object",
                    "required": [
                        "device_id"
                    ],
                    "properties": {
                        "app_version": {
                            "type": "string"
                        },
                        "device_id": {
                            "type": "string"
                        },
                        "device_type": {
                            "type": "string"
                        },
                        "os_version": {
                            "type": "string"
                        }
                    }
                },
                "phone": {
                    "type": "string"
                }
            }
        }
    }
}